hfsplus: Fix disk support for disks > 4GB
The HFS+ driver was truncating the end of the disk to 32bit implicitly
by multiplying a 32bit value with another 32bit value. Cast the first one
to 64bit, making the result 64bit wide as well.

This fixes booting from big disks for me.


git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1162 f158a5a8-5612-0410-a976-696ce0be7e32

+++ fs/hfsplus/hfsp_volume.c
@@ -207,7 +207,7 @@ volume_read_wrapper(volume * vol, hfsp_vh* vh)
                 ret = volume_readbuf(vh, p);
 		if( !ret ) {
 		    /* When reading the initial partition we must use 512 byte blocks */
-		    vol_size = vh->blocksize * vh->total_blocks;
+		    vol_size = (uint64_t)vh->blocksize * vh->total_blocks;
 		    vol->maxblocks = vol_size / HFSP_BLOCKSZ;
 		}
